Load wallet for validator
