+++ envs/.env.template
@@ -21,6 +21,10 @@ AWS_ROUTE53_HOSTED_ZONE_ID=
 
 SUBTENSOR__NETWORK=
 NETUID=
+
+
+# Wallet location
+
 # WALLET__PATH=
 WALLET__NAME=
 WALLET__HOTKEY=
