+++ envs/.env.test.template
@@ -1,14 +1,7 @@
-VALIDATOR_HOTKEY=
-VALIDATOR_PRIVATE_KEY=
-VALIDATOR_PUBLIC_KEY=
-
-
-
 # Shielded server details (EC2 instance)
 
 AWS_MINER_INSTANCE_ID=
 AWS_MINER_INSTANCE_IP=
-
 
 
 # AWS credentials
@@ -23,11 +16,27 @@ AWS_ROUTE53_HOSTED_ZONE_ID=
 AWS_ROUTE53_OTHER_HOSTED_ZONE_ID=
 
 
-
 # Bittensor credentials
 
 SUBTENSOR__NETWORK=
 NETUID=
+
+
+# Miner wallet location
+
 # WALLET__PATH=
-WALLET__NAME=
-WALLET__HOTKEY=
+WALLET__NAME=bt_ddos_shield
+WALLET__HOTKEY=miner
+
+
+# Validator wallet location
+
+# VALIDATOR_WALLET__PATH=
+VALIDATOR_WALLET__NAME=bt_ddos_shield
+VALIDATOR_WALLET__HOTKEY=validator
+
+
+# Validator encrypting key
+
+VALIDATOR_PRIVATE_KEY=
+VALIDATOR_PUBLIC_KEY=
