+++ tests/conftest.py
@@ -1,6 +1,6 @@
 import pytest
-
 from pydantic import Field
+
 from bt_ddos_shield.miner_shield import ShieldSettings
 from bt_ddos_shield.validator import ValidatorSettings
 
@@ -24,6 +24,7 @@ def shield_settings() -> ShieldTestSettings:
 
 class ValidatorTestSettings(ValidatorSettings):
     validator_public_key: str = Field(min_length=1)
+    """ Hex representation of secp256k1 public key of validator """
     miner_instance_port: int = 8080
 
     model_config = {
